fix(Goal): calculate last state comment
