+++ src/utils/db/calculatedGoalsFields.ts
@@ -1,12 +1,6 @@
-import { Activity, Reaction, Role, User, Ghost, Comment, State } from '@prisma/client';
+import { Role } from '@prisma/client';
 
 import { calcAchievedWeight } from '../recalculateCriteriaScore';
-import { ReactionsMap } from '../../types/reactions';
-import { safeGetUserName } from '../getUserName';
-
-interface UserActivity {
-    activity: Activity & { user: User; ghost: Ghost | null };
-}
 
 export const addCommonCalculatedGoalFields = (goal: any) => {
     const _shortId = `${goal.projectId}-${goal.scopeId}`;
@@ -16,52 +10,10 @@ export const addCommonCalculatedGoalFields = (goal: any) => {
             ? calcAchievedWeight(goal.goalAchieveCriteria)
             : goal.completedCriteriaWeight;
 
-    const lastCommentWithUpdateState: Comment &
-        UserActivity & { reactions: (Reaction & UserActivity)[]; state: State } = goal.comments?.[0];
-
-    let reactions: ReactionsMap = {};
-    if (lastCommentWithUpdateState) {
-        const limit = 10;
-        reactions = lastCommentWithUpdateState.reactions?.reduce<ReactionsMap>((acc, cur) => {
-            const data = {
-                activityId: cur.activityId,
-                name: safeGetUserName(cur.activity),
-            };
-
-            if (acc[cur.emoji]) {
-                acc[cur.emoji].count += 1;
-                acc[cur.emoji].authors.push(data);
-            } else {
-                acc[cur.emoji] = {
-                    count: 1,
-                    authors: [data],
-                    remains: 0,
-                };
-            }
-
-            return acc;
-        }, {});
-
-        for (const key in reactions) {
-            if (key in reactions) {
-                const { authors } = reactions[key];
-
-                if (authors.length > limit) {
-                    reactions[key].authors = authors.slice(0, limit);
-                    reactions[key].remains = authors.length - limit;
-                }
-            }
-        }
-    }
-
     return {
         _shortId,
         _hasAchievementCriteria,
         _achivedCriteriaWeight,
-        _lastComment: {
-            ...lastCommentWithUpdateState,
-            reactions,
-        },
     };
 };
